Aligned DEBUG speed to RELEASE. - Print Graphics vendor together with OpenGL version. - Removed unused projects. - Readjusted Fresnel to 0.04.

+++ source/gdevice/gpu/opengl/gl.h
@@ -81,7 +81,7 @@ namespace GL
         return (const char *)glGetString(GL_RENDERER);
     }
 
-	double version()
+	float version()
 	{
 		return atof((const char *)glGetString(GL_VERSION));
 	}
